backend2/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.30
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=2.8.0
pydantic[email]>=2.8.0
gunicorn>=21.2.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0